fix(B2B-707): uploaded_by_user_id is INT, because users.id is
The migration failed against dev on its foreign key:

  ERROR 3780 (HY000): Referencing column 'uploaded_by_user_id' and referenced
  column 'id' in foreign key constraint 'fk_qdoc_user' are incompatible

migration_quotecenter_schema.sql declares uploaded_by_user_id BIGINT to match a
users table it also declares with a BIGINT id. That users table was never
created. The live `users` is legacy's and its id is INT, so the FK is rejected
outright.

Built from SHOW COLUMNS rather than from the migration file, exactly as
repository/CLAUDE.md says to — this is the same class of gap it describes for
quotation_delivery. Quotation.userId was already an Integer for the same reason,
which is the confirmation.

The column, the entity field, the DTO and both signatures move together.

+++ api/src/main/java/com/dqs/api/controller/QuotationDocumentController.java
@@ -33,7 +33,7 @@ public class QuotationDocumentController {
             @Parameter(description = "ID de la cotización") @PathVariable Long id,
             @RequestParam("file") MultipartFile file,
             @Parameter(description = "Usuario que adjunta; opcional hasta que exista autenticación")
-            @RequestParam(value = "userId", required = false) Long userId) {
+            @RequestParam(value = "userId", required = false) Integer userId) {
         log.info("[QuotationDocumentController] POST /api/v1/quotations/{}/vouchers name={} bytes={}",
                 id, file.getOriginalFilename(), file.getSize());
         return ResponseEntity.status(HttpStatus.CREATED)
